requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
redis
openai
torch
transformers
openai-whisper
imageio-ffmpeg==0.6.0
gTTS
scikit-learn==1.6.1
soundfile
numpy
python-dotenv
rapidfuzz
pytesseract
pillow
pytest
